Add devel/gcli: CLI utility to interact with various Git forges
GCLI is a simple and portable CLI tool for interacting with GitHub,
GitLab, and Gitea from the command line.

This tool was written because the official GitHub CLI tool only supports
GitHub.  The author wanted a simple unified tool for various git forges
such as GitHub and GitLab because every forge does things differently
yet all build on Git and purposefully break with its philosophy.

Also, the official tool from Github is written in Go, which does manual
DNS resolution which is a massive security vulnerability for people
using Tor as it leaks your IP to the DNS server. This program builds
upon libcurl, which obeys the operating system's DNS resolution
mechanisms and thus also works with Tor.

PR:		267224