games/gzdoom: fix runtime against SDL newer than 2.0
The game was supposed to be restricted to newer versions of SDL than
2.0.6, but they inadvertently only checked the patch number of the
current SDL version.  The provided patch is pulled from a later version
of GZDoom and uses the correct SDL version check macro so that newer
versions of SDL don't need to get up to patchlevel=6 to work again.

MFH:	2022Q4