sysutils/check_reload_status: Ensure recvd buffers are null terminated. Fixes #13598
It appears that php-fpm can send back fastcgi packets for STDOUT and STDERR
bytestreams that do not include a terminating NULL as expected by fcgicli. If
there is no newline in the buffer, strsep() will not replace it with a
terminating NULL, resulting in possible garbage output when printing the
buffer. To ensure that all received buffers are NULL terminated and can be
treated as C strings, read_packet() is changed to allocate len+1 bytes for the
received payload, initializing the final byte to 0.