Fixes in Youtube SafeSearch
Even though there was None selected. Youtube content was getting filtered as `squid.conf` generated by `squid.inc` was adding headers with none value.

A simple fix by setting the value of None drop down item with empty string as `squid.inc` will check for a empty string before adding it to `squid.conf`. As it will receive empty value so it will not add it to `squid.conf` file. As a result youtube will not get filtered on None Selection.