net/pfSense-pkg-pfblockerNG-devel: Protect widget from indexing error strings. Fixes #13619
pfSense_get_pf_rules() builds an array of rule property arrays and returns them
to the widget code, which iterates over them and displays rule information in
the widget. However, on the chance that pfSense_get_pf_rules() gets an error
calling pfctl_get_rule(), it will add a keyed error string to the return value
and return, leaving the returned array with a scalar element that we can attempt
to index as an array. This change checks the iterated $prule to verify it is an
array to prevent this occurence.