net/pfSense-pkg-frr: Correct writing of sorted route-maps. Fixes #13642
Commit fd578ce7768513651feea89bf134ce791e810346 inadvertently introduced a bug
in frr_confg_sort_route_map() where the return value of usort() was expected to
be the sorted array of route maps, when it is in fact always the value true and
the array parameter is passed by reference and sorted in place. As a result, the
sorted and saved configuration is always empty. This change corrects the bug by
setting the value of $frr_routemap_conf to the config as is expected.