net/pfSense-pkg-pfBlockerNG-devel: Properly close handles in pfblockerng_alerts.php. Fixes #13446
The original code unconditionally closes $handle without checking its value
first, with the exception of one call which is wrapped in a conditional check of
the value which was added in 3bdda7. This change improves the closing of $handle
by only calling fclose() on it from within the conditional block checking the
corresponding fopen() return. All fclose() calls are accounted for in this
change.