net/pfSense-pkg-freeradius3: Improve validation of user params that can easily overflow: Fixes #13842
Integer attribute values sent over the line in RADIUS cannot exceed 4
octets. Values in a user's configuration that are packed into integer attributes
will be truncated to 4 bytes without any errors being logged. This problem is
mostly noticeable with the varusersmaxtotaloctets and
varusersmaxbandwith[up/down] input parameters, which are multiplied by 1024^2
and 1000 (respectively). This commit adds tighter validation of these input
parameters to prevent the administrator entering values that result in uint32
overflows in the user config.