net/pfSense-pkg-freeradius3: Remove max octets limit for users. Fixes #13947
The max octets field limit introduced in 5833ede to protect captive portal from
receiving integer overflow values of the pfSense-Max-Total-Octets attribute also
introduced and artificial hard limit in accounting for admins that are not using
captive portal or are using captive portal with periodic reauthentication
enabled. This change removes the upper limit to the field validation and
replaces it with an informative warning.