sysutils/pfSense-upgrade: Make boot loader upgrade more reliable. Fixes #10908
The pfSense-boot package utilizes a POST-INSTALL script to mount /boot/efi and
copy the newly installed efi bootloader into the efi fat. This can fail for a
number of reasons, not least of which being that /boot/efi could fail to mount
because it is already mounted. POST-INSTALL scripts do not affect the return
status of the pkg command, however, so these are non-terminal. Additionally, a
number of packages are upgraded, including the kernel, before the core meta
package that pulls the boot package as a dependency is actually upgraded. If the
boot package upgrade fails to install the current loader to the efi fat the
system may be left unbootable as the legacy efi loader may not be capable of
mounting the ZFS root.

To prevent this, pfSense-upgrade is altered to upgrade pfSense-boot early,
before pfSense-rc and pfSense-kernel, and the pkg output is captured and checked
for text indicating the failure of the POST-INSTALL script. If this is detected,
the upgrade is aborted. The upgrade of the pfSense-boot is also forced so that
even if the current version is installed the POST-INSTALL script will always be
invoked to update the efi loader.

Additionally, prior to attempting to upgrade pfSense-boot, /boot/efi is
unmounted on the chance that it is mounted either by the user or is left mounted
unintentionally by some other script. This can cause the POST-INSTALL script to
fail to install the updated loader, as it assumes /boot/efi to be an available
mountpoint.