sysutils/pfSense-upgrade: Implement distinct update/upgrade checks
Historically, pfSense-upgrade has relied on the 'stable' repo being pointed to a
new version repository to be able to detect new updates or full versions by
simply querying the configured pkg repo. This change adds facilities to
pfSense-repo-setup and pfSense-upgreade for checking for available updates on
all available repo configs to search for upgrades, allowing opt-in upgrade
checks that do not require a named repo configuration to be modified to point to
a new pair of release repositories.

In this change, pfSense-repo-setup accepts optional destination dir and repo
conf path to create a valid pkg.conf, repos directory with simlink to the
provided repo conf, and pkg database directory for an arbitrary repo
config. When accepting these optional parameters, the system-wide pkg.conf and
pfSense.conf are untouched, and pkg may be invoked with the -C parameter to use
the pkg.conf in the destination directory. This generated pkg.conf additionally
sets the pkg database directory appropriately so avoid stepping on the
system-wide repo database.

pfSense-upgrade is altered to accept a new -C flag, which is analagous to -c but
invokes the new mode of operation in pfSense-repo-setup (if necessary), to
generate a repo config and database in /tmp for other non-devel repos to search
for the highest available pfSense version beyond the system configured
repositories. This mode is somewhat slower as it must do additional pkg
rqueries, and may need to initialize a local repo database if one isn't already
in existence.

(cherry picked from commit e78ab8d2921c400f6f2ce7a7bb13cda29e19bd5b)
(cherry picked from commit b41f490c76928d78933834e305fc08328624c7a2)