sysutils/tailspin: add port: Log file highlighter
tailspin works by reading through a log file line by line, running a
series of regexes against each line. The regexes recognize patterns
like dates, numbers, severity keywords and more.

tailspin does not make any assumptions on the format or position of
the items it wants to highlight. For this reason, it requires no
configuration or setup and will work predictably regardless of the
format the log file is in.

WWW: https://github.com/bensadeh/tailspin