databases/mysq57-{client, server}: Update to latest release 5.7.43
Bugs Fixed

Group Replication: After one machine halted and restarted in a
three-node MySQL InnoDB Cluster, one node failed to start; after
restarting all nodes, the cluster shut down unexpectedly.

Our thanks to Zetang Zeng for the contribution. (Bug #34976442)

Fortified parsing of the network packet data sent by the server to
the client. (Bug #35374491)

Some floating-point literals were not always handled correctly. (Bug

Executing a query with an implicit aggregation should return exactly
one row, unless the query has a HAVING clause that filters out the row,
but a query with a HAVING clause which evaluated to FALSE sometimes
ignored this, and returned a row regardless. (Bug #14272020)

During optimization, range-select tree creation uses logic which
differs based on the left-hand side of the IN() predicate. For a field
item, each value on the right-hand side is added to an OR tree to create
the necessary expression. In the case of a row item comparison (example:
WHERE (a,b) IN ((n1,m1), (n2, m2), ...)), an expression in disjunctive
normal form (DNF) is needed. A DNF expression is created by adding an
AND tree with column values to an OR tree for each set of RHS values,
but instead the OR tree was added to the AND tree causing the tree merge
to require exponential time due to O(n2) runtime complexity. (Bug

Changelog: https://dev.mysql.com/doc/relnotes/mysql/5.7/en/news-5-7-43.html

Sponsored by:	Netzkommune GmbH