sysutils/opensbi: fix build with clang 17
Building sysutils/opensbi with clang 17 results in the following errors:

  clang: error: unsupported option '-mno-unaligned-access' for target 'riscv64-unknown-elf'
  clang: error: unsupported option '-mno-unaligned-access' for target 'riscv64-unknown-elf'
  [... many more times...]

The target-specific option -mno-unaligned-access (which is emitted from
the compiler driver when -mstrict-align is used) did not do anything for
riscv64, but clang 17 is now more strict about reporting unsupported
options. Remove the -mstrict-align option from the Makefile's CFLAGS and
ASFLAGS to fix it.

PR:		273753
MFH:		2023Q4