devel/php-pfSense-module: fix memory leak
We array_init() 'addr' (which causes PHP to allocate memory), but potentially
break out before adding 'addr' to anything.
That causes PHP to hold on to the memory until it stops running, when it frees
all of its memory. That means valgrind (or similar tools) don't actually see
this leak.

Explicitly release the allocated memory if we're not going to be using 'addr'.

PR:	https://redmine.pfsense.org/issues/15471