Update squid.inc
$sslproxy_options = "NO_SSLv3, NO_TLSv1, NO_TLSv1_1";

This variable is also used with 

tls_outgoing_options

However without the feature flags set TLS1 and TLS1_1 is being used, this can be seen inside of pcap files. This causes issues with websites sending change cipher requests. Please set this directive's feature flags to no longer use TLS1 or TLS1.1. This leaves TLS1.2 and TLS1.3 for use with the proxy. I have confirmed this forces use of TLS1.2 and TLS1.3. With this feature flag set TLS1 and 1.1 are no longer seen in pcap files when communicating with the proxy. The pcap files were generated inside of pfsense to find this issue.