net-mgmt/pfSense-pkg-arpwatch: Add missing dot between hostname and domain
I saw arpwatch notices here the hostname was missing a dot:

    [hostname=pear, domain=example.com]
    Subject: pearexample.com - Arpwatch Notification : changed ethernet address

The bug is in sendmail_proxy.php is was surprisingly resistent
to local patching. Today I tracked down that it's created at run
time by pfSense-pkg-arpwatch. I tracked it to 99368e36:

 -       $send_subject = "{$config['system']['hostname']}.{$config['system']['domain']} - Arpwatch Notification : {$subject[1]}";
 +       $send_subject = config_get_path('system/hostname') . config_get_path('system/domain') . " - Arpwatch Notification : {$subject[1]}";

This fix for this regression is to add a string with the dot.